[#14] make docker images game version
Done
1. make docker images minecraft version
