+++ game-server/dokertest/Dockerfile_mc1.16.5
@@ -1,0 +1,24 @@
+FROM alpine:3.19.1
+
+WORKDIR /server/
+
+RUN apk add openjdk11-jre
+
+COPY mc1.16.5/craftbukkit-1.16.5.jar /server/
+
+COPY control/. /control/
+
+RUN chmod +x /control/start.sh
+RUN chmod +x /control/stop.sh
+
+RUN java -jar /server/craftbukkit-1.16.5.jar 
+
+RUN sed -i 's/eula=false/eula=true/g' /server/eula.txt
+
+WORKDIR /
+
+CMD ["java", "/control/inputAndOutput.java"]
+
+
+# CMD ["java", "-Xmx2G", "-jar", "/server/craftbukkit-1.16.5.jar"]
+#
