+++ game-server/dokertest/Dockerfile_mc1.19.2
@@ -1,0 +1,23 @@
+FROM alpine:3.19.1
+
+WORKDIR /server/
+
+RUN apk add openjdk18-jre
+
+COPY mc1.19.2/craftbukkit-1.19.2.jar /server/
+
+COPY control/. /control/
+
+RUN chmod +x /control/start.sh
+RUN chmod +x /control/stop.sh
+
+RUN java -jar /server/craftbukkit-1.19.2.jar 
+
+RUN sed -i 's/eula=false/eula=true/g' /server/eula.txt
+
+WORKDIR /
+
+CMD ["java", "/control/inputAndOutput.java"]
+
+
+# CMD ["java", "-Xmx2G", "-jar", "/server/craftbukkit-1.92.2.jar"]
