+++ game-server/dokertest/Dockerfile_tester
@@ -1,0 +1,26 @@
+FROM alpine:3.19.1
+
+WORKDIR /server/
+
+
+COPY mc1.16.5/craftbukkit-1.16.5.jar /server/
+
+COPY control/. /control/
+
+RUN chmod +x /control/start.sh
+RUN chmod +x /control/stop.sh
+CMD ["sleep", "infinity"]
+# RUN apk add openjdk16-jre
+
+
+# RUN java -jar /server/craftbukkit-1.16.5.jar 
+
+# RUN sed -i 's/eula=false/eula=true/g' /server/eula.txt
+
+# WORKDIR /
+
+# CMD ["java", "/control/inputAndOutput.java"]
+
+
+# CMD ["java", "-Xmx2G", "-jar", "/server/craftbukkit-1.20.4.jar"]
+#
