+++ game-server/dokertest/control/input.txt
@@ -1,0 +1,1 @@
+ÿþ
