--- game-server/dokertest/input.txt
@@ -1,1 +1,0 @@
-ÿþ
